Feat: info-box 컴포넌트 tailwind로 변경
